Fixes lot of bugs

+++ src/dlist.c
@@ -173,9 +173,7 @@ dlist_find(dlist_t *list, const void *data)
     return NULL;
 }
 
-dlist_t*
-dlist_search(dlist_t *list,
-             int8_t (*cmp) (void *data, void *pattern), void *pattern)
+dlist_t *dlist_search( dlist_t *list, int (*cmp) (void *data, void *pattern), void *pattern)
 {
     dlist_t *p;
